+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "7-segment scanner".
 - Tests for "Lumen".
 - Tests for "A child's play".
+- Tests for "Sum of spiral's diagonals".
 
 ## [1.6.0] - 2022-03-11
 ### Added